Move data structure files to src and change naming convention
