+++ src/Node.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "node.h"
+#include "Node.h"
 
 Node* CreateNode(char* key, void* val, DataType val_type)
 {
