+++ src/Queue.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "queue.h"
-#include "node.h"
+#include "Queue.h"
+#include "Node.h"
 
 Queue* CreateQueue(void)
 {
